Validating username as author name before registration implemented.
