+++ aea/cli/init.py
@@ -28,12 +28,10 @@ from aea.cli.common import (
     AEA_LOGO,
     AUTHOR,
     Context,
-    NOT_PERMITTED_AUTHORS,
     _get_or_create_cli_config,
-    _is_permitted_author_handle,
-    _is_valid_author_handle,
     _update_cli_config,
     pass_ctx,
+    validate_author_name
 )
 from aea.cli.login import do_login
 from aea.cli.register import do_register
@@ -42,66 +40,21 @@ from aea.cli.registry.utils import check_is_author_logged_in, is_auth_token_pres
 from aea.configurations.base import PublicId
 
 
-def _local_init(author: Optional[str] = None) -> str:
-    """
-    Create an author name for local usage only.
-
-    :param author: the author name (optional)
-    """
-    is_acceptable_author = False
-    if (
-        author is not None
-        and _is_valid_author_handle(author)
-        and _is_permitted_author_handle(author)
-    ):
-        is_acceptable_author = True
-        valid_author = author
-    while not is_acceptable_author:
-        author_prompt = click.prompt(
-            "Please enter the author handle you would like to use", type=str
-        )
-        valid_author = author_prompt
-        if _is_valid_author_handle(author_prompt) and _is_permitted_author_handle(
-            author_prompt
-        ):
-            is_acceptable_author = True
-        elif not _is_valid_author_handle(author_prompt):
-            is_acceptable_author = False
-            click.echo(
-                "Not a valid author handle. Please try again. "
-                "Author handles must satisfy the following regex: {}".format(
-                    PublicId.AUTHOR_REGEX
-                )
-            )
-        elif not _is_permitted_author_handle(author_prompt):
-            is_acceptable_author = False
-            click.echo(
-                "Not a permitted author handle. The following author handles are not allowed: {}".format(
-                    NOT_PERMITTED_AUTHORS
-                )
-            )
-
-    return valid_author
-
-
-def _registry_init(author: Optional[str] = None) -> str:
+def _registry_init(username: str) -> str:
     """
     Create an author name on the registry.
 
     :param author: the author name
     """
-    if author is not None and is_auth_token_present():
-        check_is_author_logged_in(author)
+    if username is not None and is_auth_token_present():
+        check_is_author_logged_in(username)
     else:
         is_registered = click.confirm("Do you have a Registry account?")
-
         if is_registered:
-            username = click.prompt("Username", type=str)
             password = click.prompt("Password", type=str, hide_input=True)
             do_login(username, password)
         else:
             click.echo("Create a new account on the Registry now:")
-            username = click.prompt("Username", type=str)
             email = click.prompt("Email", type=str)
             password = click.prompt("Password", type=str, hide_input=True)
 
@@ -129,10 +82,10 @@ def do_init(author: str, reset: bool, registry: bool) -> None:
     """
     config = _get_or_create_cli_config()
     if reset or config.get(AUTHOR, None) is None:
+        author = validate_author_name(author)
         if registry:
-            author = _registry_init(author)
-        else:
-            author = _local_init(author)
+            author = _registry_init(username=author)
+
         _update_cli_config({AUTHOR: author})
         config.pop(AUTH_TOKEN_KEY, None)  # for security reasons
         success_msg = "AEA configurations successfully initialized: {}".format(config)
