+++ aea/cli/register.py
@@ -21,7 +21,7 @@
 
 import click
 
-from aea.cli.common import _update_cli_config
+from aea.cli.common import _update_cli_config, validate_author_name
 from aea.cli.registry.registration import register as register_new_account
 from aea.cli.registry.settings import AUTH_TOKEN_KEY
 
@@ -39,6 +39,7 @@ def do_register(
 
     :return: None
     """
+    username = validate_author_name(username)
     token = register_new_account(username, email, password, password_confirmation)
     _update_cli_config({AUTH_TOKEN_KEY: token})
     click.echo("Successfully registered and logged in: {}".format(username))
